refactor(card-browser): move 'tags dialogs' to Fragment
* showEditTagsDialog
* showFilterByTagsDialog
