+++ AnkiDroid/src/test/java/com/ichi2/anki/CardBrowserTest.kt
@@ -1630,7 +1630,7 @@ suspend fun CardBrowser.searchCardsSync(query: String) {
 }
 
 suspend fun CardBrowser.filterByTagSync(vararg tags: String) {
-    filterByTag(*tags)
+    cardBrowserFragment.filterByTag(*tags)
     viewModel.searchJob?.join()
 }
